2022/07/15 변형 PART1 Chapter8 연습문제 2번~PART1 Chapter9 연습문제 8번

+++ PART01/chapter_4_new/C_ex4-1.c
@@ -14,7 +14,7 @@ int main(){
 
 int main(){
     
-    printf("%s %.2lf \n","10.00 + 20.00 =",(double)10 + (double)20);
+    printf("%s %.2lf \n","10.00 + 20.00 =",(double)10 + 20);
     
     return 0;
 }
